move fd and path from wasi_snapshot_preview1 into Entry methods
anything that just requires an entry be in scope is moved to entry/fd.rs
and entry/path.rs, which just contain `impl Entry` blocks.

Additionally, the bodies of these functions were transformed so that as
much GuestPtr dereference as possible happens out in
wasi_snapshot_preview1.  Entry methods just get a `& T` or `&mut T`
passed in.

However, there are a few functions where multiple GuestPtrs are passed
into a wasi_snapshot_preview1 function. We cannot safely dereference
both at once, because it is semantically valid for these memory
regions to overlap, but the wiggle borrow checker will refuse to
borrow overlapping regions. In these cases, we have to pass the GuestPtr
into the Entity method to be dereferenced in a smaller scope where
exclusive access is possible.