poll_oneoff: factor part-way into sched, make error handling more clear
prior to this, we would report a missing entry for an fd as an error
event, but not a bad capability. why? i think because of the way the
code was written.

Also, we'd poll rather than report subscription errors as early as
possible. This is an odd choice because maybe none of the non-errored
subscriptions would ever come due, leaving the user waiting forever
with no error to react to. So, now poll_oneoff returns early if
any of the subscriptions were parsed as an error.

The extraction of the `Event` type and its subordinates from sched
and sys is still incomplete. That representation changes in ephemeral
and working on this refactoring has made me want to make even more
changes to the design of the interface (userdata isnt a good design
imho!) so i'm going to hold off on that for a minute till we come
up with a better consensus on the design there.